requirements.txt
annotated-types==0.7.0
anyio==4.12.0
blinker==1.9.0
certifi==2025.11.12
click==8.3.1
deprecation==2.1.0
Flask==3.0.0
Flask-Login==0.6.3
gotrue==2.12.4
gunicorn==21.2.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.27.0
hyperframe==6.1.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
markdown2==2.4.12
MarkupSafe==3.0.3
packaging==25.0
postgrest==0.18.0
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
realtime==2.27.0
six==1.17.0
sniffio==1.3.1
storage3==0.9.0
StrEnum==0.4.15
supabase==2.10.0
supafunc==0.7.0
typing-inspection==0.4.2
typing_extensions==4.15.0
websockets==15.0.1
Werkzeug==3.1.4
pandas==2.1.4
plotly==5.18.0
requests==2.32.5
markdown2==2.4.12
groq==0.9.0
Authlib==1.3.0
cryptography==42.0.5
